feat: restructure participant detail page and implement biodata management components
