+++ src/app/dashboard/participant/[id]/page.tsx
@@ -1,14 +1,15 @@
 "use client";
 
 import { Button } from "@/components/ui/button";
-import ParticipantDetailForm from "@/features/participant/form/participant-detail-form";
-import BiodataPeserta from "@/features/participant/components/biodata-peserta";
+import ParticipantDetailForm from "@/features/participant/form/detail/participant-detail-form";
+import BiodataPeserta from "@/features/participant/components/detail/biodata-peserta";
 import { EyeIcon, PrinterIcon } from "lucide-react";
 import React, { use } from "react";
 import { useQuery } from "@tanstack/react-query";
 import handleRequest from "@/axios/request";
 import { ParticipantDetail } from "@/features/participant/dto";
 import { toast } from "sonner";
+import Header from "@/features/participant/components/detail/header";
 
 const ParticipantDetailPage = ({ params }: { params: Promise<{ id: string }> }) => {
   const { id } = use(params);
@@ -50,18 +51,7 @@ const ParticipantDetailPage = ({ params }: { params: Promise<{ id: string }> }) 
   return (
     <div className="p-4 border rounded-lg bg-white space-y-6">
       <div className="">
-        <div className="flex justify-between py-4 px-1">
-          <h1 className="text-xl font-bold">Detail Peserta</h1>
-          <div className="flex gap-2">
-            <Button className="bg-gray-800 text-white">
-              <EyeIcon /> Lihat Bukti Pembayaran
-            </Button>
-            <Button className="bg-gray-800 text-white">
-              <PrinterIcon /> Cetak Kartu Peserta
-            </Button>
-          </div>
-        </div>
-
+        <Header participantId={id} />
         <ParticipantDetailForm id={id} />
       </div>
